extract helper functions to new file
